api_server/requirements.txt
quart
quart_sqlalchemy
uvicorn[standard]
aiohttp
azure-search-documents --pre
azure-storage-blob
openai
python-dotenv
bing-images
Pillow